third_party/WebKit/LayoutTests/geolocation-api/reentrant-success.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test.js"></script>
<script src="../resources/mojo-helpers.js"></script>
<script src="resources/geolocation-mock.js"></script>
</head>
<body>
<script>
description("Tests that reentrant calls to Geolocation methods from the success callback are OK.");

var mockLatitude = 51.478;
var mockLongitude = -0.166;
var mockAccuracy = 100.0;

var position;

geolocationMock.then(mock => {
    mock.setGeolocationPermission(true);
    mock.setGeolocationPosition(mockLatitude,
                                mockLongitude,
                                mockAccuracy);

    var successCallbackInvoked = false;
    navigator.geolocation.getCurrentPosition(function(p) {
        if (successCallbackInvoked) {
            testFailed('Success callback invoked unexpectedly');
            finishJSTest();
        }
        successCallbackInvoked = true;

        position = p;
        shouldBe('position.coords.latitude', 'mockLatitude');
        shouldBe('position.coords.longitude', 'mockLongitude');
        shouldBe('position.coords.accuracy', 'mockAccuracy');
        debug('');
        continueTest();
    }, function(e) {
        testFailed('Error callback invoked unexpectedly');
        finishJSTest();
    });

    function continueTest() {
        mock.setGeolocationPosition(++mockLatitude,
                                    ++mockLongitude,
                                    ++mockAccuracy);

        navigator.geolocation.getCurrentPosition(function(p) {
            position = p;
            shouldBe('position.coords.latitude', 'mockLatitude');
            shouldBe('position.coords.longitude', 'mockLongitude');
            shouldBe('position.coords.accuracy', 'mockAccuracy');
            finishJSTest();
        }, function(e) {
            testFailed('Error callback invoked unexpectedly');
            finishJSTest();
        });
    }
});

window.jsTestIsAsync = true;
</script>
</body>
</html>
